Complete system health monitoring, fix template validation, add 8 superadmin management forms
- Implemented System Health & Performance Monitoring dashboard
- Created 8 new superadmin management forms:
  * Security Configuration (14 configurable settings)
  * Account Deletion (soft delete with reasons)
  * Account Activation/Deactivation
  * Activity Log Viewer (search/filter 500 logs)
  * IP Address Management (CRUD whitelist)
  * Active Session Management
  * Permission Assignment (role-based access)
  * Recycle Bin Enhancement (deleted users view)
- Fixed Jinja2 template syntax validation errors
- Added .vscode/settings.json and .htmlhintrc for proper linting
- All routes tested and functional
- Database schema updated with new tables

+++ routes/student_operations.py
@@ -151,7 +151,7 @@ def attendance(student_id):
                        present_days, absent_days, leave_days,
                        CASE 
                            WHEN (present_days + absent_days) > 0 
-                           THEN ROUND(present_days * 100.0 / (present_days + absent_days), 2)
+                           THEN ROUND((present_days * 100.0 / (present_days + absent_days))::numeric, 2)
                            ELSE 0
                        END AS percentage,
                        remarks
